feat(#42): 🎨 Change default body font to Signika
From Lodrina Solid. This font was not best for body uses so we ran a survey with the Com team and picked Signika that is highly legible and slighly adapted to our entertainement sprit.

Fix the previous commit where the file was not properly saved due to an IDE issue.